+++ order-service/database/db.go
@@ -5,7 +5,7 @@ import "order-service/model"
 type Database interface {
 	GetAllOrders() ([]model.Order, error)
 	InsertOrder(or model.OrderRequest) error
-	GetOrdersByUser(userId uint64) ([]model.Order, error)
+	GetOrdersByUser(userId int32) ([]model.Order, error)
 	Close()
 }
